+++ code/providers/virtualbox/configure_instance.rb
@@ -7,6 +7,7 @@ def configure_instance_virtualbox(machine, instance_profile, provider = 'virtual
   configure_vagrant_box(machine, instance_profile, provider)
   configure_instance_hardware(machine, instance_profile, provider)
   configure_networking(machine, instance_profile, provider)
+  configure_vagrant_dns(machine, instance_profile, provider)
   configure_communication(machine, instance_profile, provider)
   configure_filesystems(machine, instance_profile, provider)
   execute_system_commands(machine, instance_profile, provider)